+++ terra-jupyter-bioconductor/CHANGELOG.md
@@ -1,3 +1,10 @@
+## 2.2.4 - 2024-04-23
+- Update `terra-jupyter-r` to `2.2.4`
+  - Update `terra-jupyter-base` to `1.1.4`
+    - Change welder user id
+
+Image URL: `us.gcr.io/broad-dsp-gcr-public/terra-jupyter-bioconductor:2.2.4`
+
 ## 2.2.3 - 2023-10-03
 - Update `terra-jupyter-r` to `2.2.3`
     - Update `terra-jupyter-base` to `1.1.3`