+++ terra-jupyter-hail/CHANGELOG.md
@@ -1,3 +1,10 @@
+## 1.1.8 - 2024-04-23
+- Update `terra-jupyter-python` to `1.1.5`
+  - Update `terra-jupyter-base` to `1.1.4`
+    - Change the welder user id
+
+Image URL: `us.gcr.io/broad-dsp-gcr-public/terra-jupyter-hail:1.1.8`
+
 ## 1.1.7 - 2023-11-01
 - Pin `gnomad` to `0.6.4`
   - See https://github.com/broadinstitute/gnomad_methods/releases for details