+++ terra-jupyter-r/CHANGELOG.md
@@ -1,4 +1,10 @@
-## 1.1.3 - 2023-10-03
+## 2.2.4 - 2024-04-23
+- Update `terra-jupyter-base` to `1.1.4`
+  - Change the welder user id
+
+Image URL: `us.gcr.io/broad-dsp-gcr-public/terra-jupyter-r:2.2.4`
+
+## 2.2.3 - 2023-10-03
 - Update `terra-jupyter-base` to `1.1.3`
    - Downgrade notebook to `6.5.4` and jupyterlab-server to `2.23.0`